Prepared terrain to start the new NO-BRAIN combat system.

+++ Projects/Remizione/Actors/Archetypes/LurkerArchetype.cs
@@ -23,6 +23,6 @@ namespace Remizione
         public override Ratio FleeChance => 0.15f;
 
         // MeleeRange
-        public override int MeleeRange => 30;
+        public override float MeleeRange => 30;
     }
 }
